requirements.txt
paho-mqtt==2.1.0
icecream==2.1.4
requests==2.32.3